Drop ignored Travis testing; adds nothing compared with Wikimedia CI
This repo hasn't been passing on Travis for a few months, and no-one
noticed, because reviewing Travis results post-merge isn't routinely
done. This doesn't add any advantage nowadays, since we test all the
applicable versions of PHP in parallel in Wikimedia CI (which wasn't
the case in 2015 when this was added for T90264).

Change-Id: Iec0f068234929a0016ee87217110deed485d25ca