build: Exclude benchmark scripts when exporting package
This was originally added via Ib85fe09. Since the scripts/ folder
was not excluded this left three scripts behind:

generate.php is only partially usable (only with the --fetch option)
when the data/ directory is missing.

The two benchmarks (benchmark.php and memstress.php) are usable.
But why?

It feels like these are all pure dev-only scripts that are never
run as part of a production environment. We have similar benchmark
scripts in other codebases and also exclude them, as far as I know.

Change-Id: Ifb833d76771716bf89141c31fd5d76fed7449211